+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 ## unreleased
 
-*   New opaque error type with more uniform, richer error messages. No more
-    `failure` dependency.
+*   BREAKING CHANGE: New opaque error type with more uniform, richer error
+    messages. No more `failure` dependency.
+*   BREAKING CHANGE: `retina::client::Stream::parameters` now returns parameters
+    by value. This allows shrinking depacketizer types.
+*   BREAKING CHANGE: `retina::codec::VideoFrame::new_parameters` is now boxed.
+    This allows shrinking `VideoFrame` and `CodecItem` by 80 bytes each (on
+    64-bit platforms). The box is only rarely populated.
 
 ## v0.0.4 (2021-06-28)